requirements.txt
requests>=2.31.0
pandas>=2.0.0
spacy>=3.7.0
sentence-transformers>=2.2.0
numpy>=1.24.0
pyarrow>=14.0.0
python-dotenv>=1.0.0

